rework launch files for hydra system
This commit adds launch files to facilitate launching namespaced hardware interfaces individually on each robot computer or simultaneously on a single computer for simulation.
It also adds the hydra_hardware package with a simulation ros_control_boilerplate hardware interface for the positioner. This will be replaced with the hardware interface for the physical positioner when practical.